feat: Install script using brew whenever present

+++ scripts/deploy-airflow-minikube.sh
@@ -11,10 +11,13 @@ BASEDIR=$(cd "$(dirname "$_MY_SCRIPT")" && pwd)
 
 if [[ ! $(which helm) ]]; then
   echo Installing helm which is required for deploying to kubernetes
-  curl -LO https://git.io/get_helm.sh
-  chmod 700 get_helm.sh
-  ./get_helm.sh
-  rm get_helm.sh;
+  if [[ $(which brew) ]]; then
+    brew install kubernetes-helm; else
+    curl -LO https://git.io/get_helm.sh
+    chmod 700 get_helm.sh
+    ./get_helm.sh
+    rm get_helm.sh;
+  fi
 fi
 
 # Apply the tiller service account on the minikube cluster
